+++ packages/marketplace-registry/src/registry.ts
@@ -5,8 +5,9 @@ export const registry: MarketplaceManifest[] = [
 	{
 		id: "app.inlang.cli",
 		icon: "https://inlang.com/favicon/safari-pinned-tab.svg",
-		coverImage:
+		gallery: [
 			"https://cdn.jsdelivr.net/gh/inlang/monorepo@latest/inlang/assets/marketplace/cli-marketplace-cover.jpg",
+		],
 		displayName: {
 			en: "CLI",
 		},
@@ -24,8 +25,9 @@ export const registry: MarketplaceManifest[] = [
 	{
 		id: "app.inlang.editor",
 		icon: "https://inlang.com/favicon/safari-pinned-tab.svg",
-		coverImage:
+		gallery: [
 			"https://cdn.jsdelivr.net/gh/inlang/monorepo@latest/inlang/assets/marketplace/editor-marketplace-cover.jpg",
+		],
 		displayName: {
 			en: "Web Editor",
 		},
@@ -44,8 +46,9 @@ export const registry: MarketplaceManifest[] = [
 	{
 		id: "app.inlang.ideExtension",
 		icon: "https://inlang.com/favicon/safari-pinned-tab.svg",
-		coverImage:
+		gallery: [
 			"https://cdn.jsdelivr.net/gh/inlang/monorepo@latest/inlang/assets/marketplace/ide_extension-marketplace-cover.jpg",
+		],
 		displayName: {
 			en: "IDE Extension",
 		},
@@ -53,7 +56,7 @@ export const registry: MarketplaceManifest[] = [
 			en: "This extension allows you to use inlang in your IDE, enabling translating content directly in your code.",
 		},
 		readme: {
-			en: "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/ide-extension/README.md",
+			en: "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/ide-extension/MARKETPLACE.md",
 		},
 		keywords: ["ide", "extension", "vscode"],
 		publisherName: "inlang",
@@ -82,7 +85,7 @@ export const registry: MarketplaceManifest[] = [
 		id: "library.inlang.paraglideJsSveltekit",
 		icon: "https://emojis.wiki/thumbs/emojis/parachute.webp",
 		displayName: {
-			en: "paraglideJS",
+			en: "paraglideJS (former SDK-JS)",
 		},
 		description: {
 			en: "A fully configurable JavaScript library that integrates within your framework.",
@@ -90,7 +93,7 @@ export const registry: MarketplaceManifest[] = [
 		readme: {
 			en: "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/paraglide-js-sveltekit/README.md",
 		},
-		keywords: ["sdk", "paraglide"],
+		keywords: ["sdk", "paraglide", "sdk-js", "sveltekit", "svelte"],
 		publisherName: "inlang",
 		publisherIcon: "https://inlang.com/favicon/safari-pinned-tab.svg",
 		license: "Apache-2.0",
@@ -116,8 +119,9 @@ export const registry: MarketplaceManifest[] = [
 		$schema: "https://inlang.com/schema/marketplace-manifest",
 		id: "messageLintRule.inlang.emptyPattern",
 		icon: "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/message-lint-rules/emptyPattern/assets/icon.png",
-		coverImage:
+		gallery: [
 			"https://cdn.jsdelivr.net/gh/inlang/monorepo@latest/inlang/assets/marketplace/empty_pattern-marketplace-cover.jpg",
+		],
 		displayName: {
 			en: "Empty pattern",
 		},
@@ -138,8 +142,9 @@ export const registry: MarketplaceManifest[] = [
 		$schema: "https://inlang.com/schema/marketplace-manifest",
 		id: "messageLintRule.inlang.identicalPattern",
 		icon: "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/message-lint-rules/identicalPattern/assets/icon.png",
-		coverImage:
+		gallery: [
 			"https://cdn.jsdelivr.net/gh/inlang/monorepo@latest/inlang/assets/marketplace/identical_pattern-marketplace-cover.jpg",
+		],
 		displayName: {
 			en: "Identical pattern",
 		},
@@ -160,8 +165,9 @@ export const registry: MarketplaceManifest[] = [
 		$schema: "https://inlang.com/schema/marketplace-manifest",
 		id: "messageLintRule.inlang.messageWithoutSource",
 		icon: "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/message-lint-rules/messageWithoutSource/assets/icon.png",
-		coverImage:
+		gallery: [
 			"https://cdn.jsdelivr.net/gh/inlang/monorepo@latest/inlang/assets/marketplace/without_source-marketplace-cover.jpg",
+		],
 		displayName: {
 			en: "Message without source",
 		},
@@ -182,8 +188,9 @@ export const registry: MarketplaceManifest[] = [
 		$schema: "https://inlang.com/schema/marketplace-manifest",
 		id: "messageLintRule.inlang.missingTranslation",
 		icon: "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/message-lint-rules/missingTranslation/assets/icon.png",
-		coverImage:
+		gallery: [
 			"https://cdn.jsdelivr.net/gh/inlang/monorepo@latest/inlang/assets/marketplace/missing_translation-marketplace-cover.jpg",
+		],
 		displayName: {
 			en: "Missing translation",
 		},