+++ packages/message-lint-rules/missingTranslation/src/missingTranslation.test.ts
@@ -76,7 +76,7 @@ test("should report if no variants are defined", async () => {
 	expect(result.data[0]!.languageTag).toBe("fr")
 	expect(result.data[0]!.messageId).toBe(message1.id)
 	expect(
-		typeof result.data[0]!.body === "object" ? result.data[0]!.body.en : result.data[0]!.body,
+		typeof result.data[0]!.body === "object" ? result.data[0]!.body.en : result.data[0]!.body
 	).toContain(message1.id)
 })