+++ packages/plugins/json/src/plugin.test.ts
@@ -183,7 +183,7 @@ describe("loadMessage", () => {
 		const sourceLanguageTag = "en"
 
 		expect(
-			plugin.loadMessages!({ languageTags, sourceLanguageTag, settings, nodeishFs: fs }),
+			plugin.loadMessages!({ languageTags, sourceLanguageTag, settings, nodeishFs: fs })
 		).resolves.toBeTruthy()
 	})
 
@@ -197,7 +197,7 @@ describe("loadMessage", () => {
 
 		const languageTags = ["en", "de"]
 		expect(
-			plugin.loadMessages!({ languageTags, sourceLanguageTag, settings, nodeishFs: fs }),
+			plugin.loadMessages!({ languageTags, sourceLanguageTag, settings, nodeishFs: fs })
 		).resolves.toBeTruthy()
 	})
 
@@ -257,7 +257,7 @@ describe("loadMessage", () => {
 		const sourceLanguageTag = "en"
 
 		expect(
-			plugin.loadMessages!({ languageTags, sourceLanguageTag, settings, nodeishFs: fs }),
+			plugin.loadMessages!({ languageTags, sourceLanguageTag, settings, nodeishFs: fs })
 		).resolves.toBeTruthy()
 	})
 
@@ -274,7 +274,7 @@ describe("loadMessage", () => {
 
 		const languageTags = ["en", "de"]
 		expect(
-			plugin.loadMessages!({ languageTags, sourceLanguageTag, settings, nodeishFs: fs }),
+			plugin.loadMessages!({ languageTags, sourceLanguageTag, settings, nodeishFs: fs })
 		).resolves.toBeTruthy()
 	})
 
@@ -478,10 +478,10 @@ describe("variable reference", () => {
 		})
 		expect(getVariant(messages[0]!, { where: { languageTag: "en" } })?.pattern.length).toBe(2)
 		expect(getVariant(messages[0]!, { where: { languageTag: "en" } })?.pattern[0]?.type).toBe(
-			"Text",
+			"Text"
 		)
 		expect(getVariant(messages[0]!, { where: { languageTag: "en" } })?.pattern[1]?.type).toBe(
-			"VariableReference",
+			"VariableReference"
 		)
 	})
 
@@ -501,10 +501,10 @@ describe("variable reference", () => {
 		})
 		expect(getVariant(messages[0]!, { where: { languageTag: "en" } })?.pattern.length).toBe(2)
 		expect(getVariant(messages[0]!, { where: { languageTag: "en" } })?.pattern[0]?.type).toBe(
-			"VariableReference",
+			"VariableReference"
 		)
 		expect(getVariant(messages[0]!, { where: { languageTag: "en" } })?.pattern[1]?.type).toBe(
-			"Text",
+			"Text"
 		)
 	})
 
@@ -524,10 +524,10 @@ describe("variable reference", () => {
 			nodeishFs: fs,
 		})
 		expect(getVariant(messages[0]!, { where: { languageTag: "en" } })?.pattern[0]?.type).toBe(
-			"Text",
+			"Text"
 		)
 		expect(getVariant(messages[0]!, { where: { languageTag: "en" } })?.pattern[1]?.type).toBe(
-			"VariableReference",
+			"VariableReference"
 		)
 	})
 })
@@ -756,7 +756,7 @@ describe("formatting", () => {
 				},
 			},
 			undefined,
-			2,
+			2
 		)
 
 		const withoutNesting = JSON.stringify(
@@ -764,7 +764,7 @@ describe("formatting", () => {
 				"test.test": "test",
 			},
 			undefined,
-			4,
+			4
 		)
 
 		const fs = createNodeishMemoryFs()
@@ -923,7 +923,7 @@ describe("roundTrip", () => {
 				test: 'Single "quote" test',
 			},
 			undefined,
-			4,
+			4
 		)
 		const fs = createNodeishMemoryFs()
 		await fs.writeFile("./en.json", complexContent)