+++ packages/website/src/pages/documentation/index.page.tsx
@@ -1,78 +1,65 @@
 import { For, Show, createEffect, createSignal, onMount } from "solid-js"
 import { Layout as RootLayout } from "#src/pages/Layout.jsx"
-import { Markdown, parseMarkdown } from "#src/services/markdown/index.js"
-import type { ProcessedTableOfContents } from "./index.page.server.jsx"
 import { currentPageContext } from "#src/renderer/state.js"
-import { Callout } from "#src/services/markdown/src/tags/Callout.jsx"
 import type SlDetails from "@shoelace-style/shoelace/dist/components/details/details.js"
 import { Meta, Title } from "@solidjs/meta"
 import { Feedback } from "./Feedback.jsx"
 import { EditButton } from "./EditButton.jsx"
 import { defaultLanguage } from "#src/renderer/_default.page.route.js"
 import { useI18n } from "@solid-primitives/i18n"
+import "@inlang/markdown/css"
+import "@inlang/markdown/custom-elements"
 import tableOfContents from "../../../../../documentation/tableOfContents.json"
 
-/**
- * The page props are undefined if an error occurred during parsing of the markdown.
- */
 export type PageProps = {
-	processedTableOfContents: ProcessedTableOfContents
-	markdown: Awaited<ReturnType<typeof parseMarkdown>>
+	markdown: Awaited<ReturnType<any>>
 }
 
 export function Page(props: PageProps) {
 	let mobileDetailMenu: SlDetails | undefined
 	const [editLink, setEditLink] = createSignal<string | undefined>("")
+	const [, { locale }] = useI18n()
+
+	const getLocale = () => {
+		const language = locale() ?? defaultLanguage
+		return language !== defaultLanguage ? "/" + language : ""
+	}
 
 	createEffect(() => {
-		if (props.markdown && props.markdown.frontmatter) {
-			const markdownHref = props.markdown.frontmatter.href
-
-			const files: Record<string, string[]> = {}
-			for (const [category, documentsArray] of Object.entries(tableOfContents)) {
-				const rawPaths = documentsArray.map((document) => document)
-				files[category] = rawPaths
-			}
-
-			for (const section of Object.keys(props.processedTableOfContents)) {
-				const documents = props.processedTableOfContents[section]
-
-				if (documents) {
-					for (const document of documents) {
-						if (document.frontmatter && document.frontmatter.href === markdownHref) {
-							const index = documents.indexOf(document)
-							const fileSource = files[section]?.[index] || undefined
-
-							const gitHubLink =
-								"https://github.com/inlang/monorepo/edit/main/inlang/documentation" +
-								"/" +
-								fileSource
-
-							setEditLink(gitHubLink)
-						}
-					}
-				}
-			}
+		if (currentPageContext) {
+			setEditLink(
+				"https://github.com/inlang/monorepo/edit/main/inlang/documentation" +
+					"/" +
+					findPageBySlug(
+						currentPageContext.urlParsed.pathname
+							.replace(getLocale(), "")
+							.replace("/documentation/", "")
+					)?.path
+			)
 		}
 	})
 
-	const h2Headlines = () => {
-		const result: string[] = []
-		// @ts-expect-error - some type mismatch in the markdown parser
-		for (const child of props.markdown.renderableTree.children ?? []) {
-			// only render h2 as sub headlines
-			if (child.name === "Heading" && child.attributes.level === 2) {
-				result.push(child.children[0])
-			}
-		}
-		return result
-	}
-
 	return (
 		<>
-			{/* frontmatter is undefined on first client side nav  */}
-			<Title>{props.markdown?.frontmatter?.title}</Title>
-			<Meta name="description" content={props.markdown?.frontmatter?.description} />
+			<Title>
+				{
+					findPageBySlug(
+						currentPageContext.urlParsed.pathname
+							.replace(getLocale(), "")
+							.replace("/documentation/", "")
+					)?.title
+				}
+			</Title>
+			<Meta
+				name="description"
+				content={
+					findPageBySlug(
+						currentPageContext.urlParsed.pathname
+							.replace(getLocale(), "")
+							.replace("/documentation/", "")
+					)?.description
+				}
+			/>
 			<Meta name="og:image" content="/images/inlang-social-image.jpg" />
 			<RootLayout>
 				{/* important: the responsive breakpoints must align throughout the markup! */}
@@ -88,36 +75,56 @@ export function Page(props: PageProps) {
 							 * filteredTableContents is not available on the client.
 							 */}
 							<div class="py-14 pr-8">
-								<Show when={props.processedTableOfContents}>
-									<NavbarCommon {...props} h2Headlines={h2Headlines()} />
+								<Show when={tableOfContents && props.markdown}>
+									<NavbarCommon
+										tableOfContents={tableOfContents}
+										getLocale={getLocale}
+										headings={props.markdown
+											.match(/<h[1-3].*?>(.*?)<\/h[1-3]>/g)
+											.map((heading: string) => {
+												// We have to use DOMParser to parse the heading string to a HTML element
+												const parser = new DOMParser()
+												const doc = parser.parseFromString(heading, "text/html")
+												const node = doc.body.firstChild as HTMLElement
+
+												return node.innerText.replace(/(<([^>]+)>)/gi, "").toString()
+											})}
+									/>
 								</Show>
 							</div>
 						</nav>
 					</div>
 					{/* Mobile navbar */}
-					<nav class="fixed min-w-full z-10 -translate-x-4 sm:-translate-x-10 sm:px-6 md:hidden overflow-y-scroll overflow-auto backdrop-blur-sm">
+					<nav class="sticky top-[69px] w-screen z-10 -translate-x-4 sm:-translate-x-10 md:hidden overflow-y-scroll overflow-auto backdrop-blur-sm">
 						<sl-details ref={mobileDetailMenu}>
-							<h3 slot="summary" class="font-medium">
+							<h3 slot="summary" class="font-medium sm:pl-6">
 								Menu
 							</h3>
 							{/* `Show` is a hotfix when client side rendering loaded this page
 							 * filteredTableContents is not available on the client.
 							 */}
-							<Show when={props.processedTableOfContents}>
+							<Show when={tableOfContents && props.markdown}>
 								<NavbarCommon
-									{...props}
-									h2Headlines={h2Headlines()}
+									tableOfContents={tableOfContents}
 									onLinkClick={() => {
 										mobileDetailMenu?.hide()
 									}}
+									getLocale={getLocale}
+									headings={props.markdown
+										.match(/<h[1-3].*?>(.*?)<\/h[1-3]>/g)
+										.map((heading: string) => {
+											// We have to use DOMParser to parse the heading string to a HTML element
+											const parser = new DOMParser()
+											const doc = parser.parseFromString(heading, "text/html")
+											const node = doc.body.firstChild as HTMLElement
+
+											return node.innerText.replace(/(<([^>]+)>)/gi, "").toString()
+										})}
 								/>
 							</Show>
 						</sl-details>
 					</nav>
-					<Show
-						when={props.markdown?.renderableTree}
-						fallback={<p class="text-danger">{props.markdown?.error}</p>}
-					>
+					<Show when={props.markdown} fallback={<p class="text-danger">{props.markdown?.error}</p>}>
 						{/* 
             rendering on the website is broken due to relative paths and 
             the escaping of html. it is better to show the RFC's on the website
@@ -125,20 +132,20 @@ export function Page(props: PageProps) {
           */}
 						<div class="w-full justify-self-center mb-8 md:p-6 md:col-span-3">
 							<Show when={currentPageContext.urlParsed.pathname.includes("rfc")}>
-								<Callout variant="warning">
+								{/* <Callout variant="warning">
 									<p>
 										The rendering of RFCs on the website might be broken.{" "}
 										<a href="https://github.com/inlang/inlang/tree/main/rfcs" target="_blank">
 											Read the RFC on GitHub instead.
 										</a>
 									</p>
-								</Callout>
+								</Callout> */}
 							</Show>
 							<div
 								// change the col-span to 2 if a right side nav bar should be rendered
 								class="w-full justify-self-center md:col-span-3"
 							>
-								<Markdown renderableTree={props.markdown.renderableTree!} />
+								<Markdown markdown={props.markdown} />
 								<EditButton href={editLink()} />
 								<Feedback />
 							</div>
@@ -150,124 +157,137 @@ export function Page(props: PageProps) {
 	)
 }
 
+function Markdown(props: { markdown: string }) {
+	return (
+		<article
+			// eslint-disable-next-line solid/no-innerhtml
+			innerHTML={props.markdown}
+		/>
+	)
+}
+
 function NavbarCommon(props: {
-	processedTableOfContents: PageProps["processedTableOfContents"]
-	h2Headlines: string[]
+	tableOfContents: typeof tableOfContents
+	headings: string[]
 	onLinkClick?: () => void
+	getLocale: () => string
 }) {
 	const [highlightedAnchor, setHighlightedAnchor] = createSignal<string | undefined>("")
-	const [, { locale }] = useI18n()
-
-	const getLocale = () => {
-		const language = locale() ?? defaultLanguage
-		return language !== defaultLanguage ? "/" + language : ""
+	const replaceChars = (str: string) => {
+		return str
+			.replaceAll(" ", "-")
+			.replaceAll("/", "")
+			.replace("#", "")
+			.replaceAll("(", "")
+			.replaceAll(")", "")
+			.replaceAll("?", "")
+			.replaceAll(".", "")
 	}
 
-	const isSelected = (href: string) => {
-		if (href === currentPageContext.urlParsed.pathname.replace(getLocale(), "")) {
+	const isSelected = (slug: string) => {
+		if (
+			`/documentation/${slug}` ===
+				currentPageContext.urlParsed.pathname.replace(props.getLocale(), "") ||
+			`/documentation/${slug}` ===
+				currentPageContext.urlParsed.pathname.replace(props.getLocale(), "") + "/"
+		) {
 			return true
 		} else {
 			return false
 		}
 	}
 
-	const onAnchorClick = (anchor: string) => {
-		setHighlightedAnchor(anchor)
+	const scrollToAnchor = (anchor: string, behavior?: ScrollBehavior) => {
+		const element = document.getElementById(anchor)
+		if (element && window) {
+			window.scrollTo({
+				top: element.offsetTop - 96,
+				behavior: behavior ?? "instant",
+			})
+		}
+		window.history.pushState({}, "", `${currentPageContext.urlParsed.pathname}#${anchor}`)
 	}
 
-	onMount(() => {
-		if (
-			currentPageContext.urlParsed.hash &&
-			props.h2Headlines
-				.toString()
-				.toLowerCase()
-				.replaceAll(" ", "-")
-				// @ts-expect-error - fix after refactoring
-				.includes(currentPageContext.urlParsed.hash?.replace("#", ""))
-		) {
-			// @ts-expect-error - fix after refactoring
-			setHighlightedAnchor(currentPageContext.urlParsed.hash?.replace("#", ""))
+	const onAnchorClick = async (anchor: string) => {
+		setHighlightedAnchor(anchor)
+		scrollToAnchor(anchor)
+	}
 
-			const targetElement = document.getElementById(
-				// @ts-expect-error - fix after refactoring
-				currentPageContext.urlParsed.hash?.replace("#", ""),
-			)
+	onMount(async () => {
+		for (const heading of props.headings) {
+			if (
+				currentPageContext.urlParsed.hash?.replace("#", "").toString() ===
+				replaceChars(heading.toString().toLowerCase())
+			) {
+				/* Wait for all images to load before scrolling to anchor */
+				await Promise.all(
+					[...document.querySelectorAll("img")].map((img) =>
+						img.complete
+							? Promise.resolve()
+							: new Promise((resolve) => img.addEventListener("load", resolve))
+					)
+				)
 
-			checkLoadedImgs(() => {
-				const elementRect = targetElement!.getBoundingClientRect()
-				const offsetPosition = elementRect.top - 96 // The offset because of the fixed navbar
-
-				window.scrollBy({
-					top: offsetPosition,
-				})
-			})
+				setHighlightedAnchor(replaceChars(heading.toString().toLowerCase()))
+				scrollToAnchor(replaceChars(heading.toString().toLowerCase()), "smooth")
+			}
 		}
 	})
 
 	return (
-		<ul role="list" class="w-full space-y-3">
-			<For each={Object.keys(props.processedTableOfContents)}>
-				{(section) => (
-					<li class="">
-						<h2 class="tracking-wide pt-2 text font-semibold text-on-surface pb-2">{section}</h2>
+		<ul role="list" class="w-full space-y-3 sm:pl-6">
+			<For each={Object.keys(props.tableOfContents)}>
+				{(category) => (
+					<li>
+						<h2 class="tracking-wide pt-2 text font-semibold text-on-surface pb-2">{category}</h2>
 						<ul class="space-y-2" role="list">
-							<For
-								each={
-									props.processedTableOfContents[
-										section as keyof typeof props.processedTableOfContents
-									]
-								}
-							>
-								{(document) => (
-									<li>
-										<a
-											onClick={props.onLinkClick}
-											class={
-												(isSelected(document.frontmatter.href)
-													? "text-primary font-semibold "
-													: "text-info/80 hover:text-on-background ") +
-												"tracking-wide text-sm block w-full font-normal"
-											}
-											href={getLocale() + document.frontmatter.href}
-										>
-											{document.frontmatter.title}
-										</a>
-										<Show
-											when={props.h2Headlines.length > 0 && isSelected(document.frontmatter.href)}
-										>
-											<ul class="my-2">
-												<For each={props.h2Headlines}>
+							<For each={props.tableOfContents[category as keyof typeof props.tableOfContents]}>
+								{(page) => {
+									const slug = page.slug
+									return (
+										<li>
+											<a
+												onClick={props.onLinkClick}
+												class={
+													(isSelected(slug)
+														? "text-primary font-semibold "
+														: "text-info/80 hover:text-on-background ") +
+													"tracking-wide text-sm block w-full font-normal mb-2"
+												}
+												href={props.getLocale() + `/documentation/${slug}`}
+											>
+												{page.title}
+											</a>
+											<Show when={props.headings && props.headings.length > 0 && isSelected(slug)}>
+												<For each={props.headings}>
 													{(heading) => (
-														<li>
-															<a
-																onClick={() => {
-																	onAnchorClick(
-																		heading.toString().toLowerCase().replaceAll(" ", "-"),
-																	)
-																	props.onLinkClick?.()
-																}}
-																class={
-																	"text-sm tracking-widem block w-full border-l pl-3 py-1 hover:border-l-info/80 " +
-																	(highlightedAnchor() ===
-																	heading.toString().toLowerCase().replaceAll(" ", "-")
-																		? "font-medium text-on-background border-l-text-on-background "
-																		: "text-info/80 hover:text-on-background font-normal border-l-info/20 ")
-																}
-																href={`#${heading
-																	.toString()
-																	.toLowerCase()
-																	.replaceAll(" ", "-")
-																	.replaceAll("/", "")}`}
-															>
-																{heading}
-															</a>
-														</li>
+														<Show when={!heading.includes(page.title)}>
+															<li>
+																<a
+																	onClick={(e) => {
+																		e.preventDefault()
+																		onAnchorClick(replaceChars(heading.toString().toLowerCase()))
+																		props.onLinkClick?.()
+																	}}
+																	class={
+																		"text-sm tracking-widem block w-full border-l pl-3 py-1 hover:border-l-info/80 " +
+																		(highlightedAnchor() ===
+																		replaceChars(heading.toString().toLowerCase())
+																			? "font-medium text-on-background border-l-on-background "
+																			: "text-info/80 hover:text-on-background font-normal border-l-info/20 ")
+																	}
+																	href={`#${replaceChars(heading.toString().toLowerCase())}`}
+																>
+																	{heading.replace("#", "")}
+																</a>
+															</li>
+														</Show>
 													)}
 												</For>
-											</ul>
-										</Show>
-									</li>
-								)}
+											</Show>
+										</li>
+									)
+								}}
 							</For>
 						</ul>
 					</li>
@@ -277,28 +297,13 @@ function NavbarCommon(props: {
 	)
 }
 
-function checkLoadedImgs(anchorScroll: () => void) {
-	let imgElementsLoaded = 0
-	const imgElements = document.querySelectorAll("img")
-	const imgElementsLength = imgElements.length
-
-	if (imgElementsLength === 0) {
-		anchorScroll()
-	} else {
-		for (const img of imgElements) {
-			if (img.complete) {
-				imgElementsLoaded++
-				if (imgElementsLoaded === imgElementsLength) {
-					anchorScroll()
-				}
-			} else {
-				img.addEventListener("load", () => {
-					imgElementsLoaded++
-					if (imgElementsLoaded === imgElementsLength) {
-						anchorScroll()
-					}
-				})
+function findPageBySlug(slug: string) {
+	for (const [, pageArray] of Object.entries(tableOfContents)) {
+		for (const page of pageArray) {
+			if (page.slug === slug || page.slug === slug.replace("/documentation", "")) {
+				return page
 			}
 		}
 	}
+	return undefined
 }