+++ packages/website/src/services/markdown/index.ts
@@ -1,9 +1,0 @@
-/**
- * --------------------------------
- * This index file exports code that is supposed
- * to be used outside of this directory.
- * --------------------------------
- */
-
-export { parseMarkdown } from "./src/parseMarkdown.js"
-export { Markdown } from "./src/Markdown.jsx"